MODFEE-79 refactoring according to code review
