+++ ramls/accounts.raml
@@ -72,3 +72,8 @@ resourceTypes:
             body:
               application/json:
                 type: accountsCheckResponse
+          500:
+            description: "Internal server error"
+            body:
+              text/plain:
+                example: "Internal server error"
