+++ ramls/examples/accountsCheckRequest.sample
@@ -1,3 +1,3 @@
 {
-	"amount":"8.00"
+	"amount":8.00
 }
